✨ 투어조회기능에 일정조회, 카테고리조회 추가
.

+++ NativeNavs-BackEnd/src/main/java/com/nativenavs/tour/dto/PlanDTO.java
@@ -2,6 +2,8 @@ package com.nativenavs.tour.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -12,7 +14,7 @@ public class PlanDTO {
     private String field;         // 일정을 설명할 필드
     private String description;  // 일정 설명
     private String image;        // 일정 사진
-    private double latitude;     // 위도
-    private double longitude;    // 경도
+    private BigDecimal latitude;     // 위도
+    private BigDecimal longitude;    // 경도
     private String addressFull;  // 주소
 }
